Revert "Open 200 - Tabbed Open dialog (#1179)"
This reverts commit e9d04f18e7c383401df837b80e46f1002f85ed69.
