+++ src/Components/Open/OpenModelControl.jsx
@@ -6,7 +6,7 @@ import useStore from '../../store/useStore'
 import {ControlButtonWithHashState} from '../Buttons'
 import OpenModelDialog from './OpenModelDialog'
 import {HASH_PREFIX_OPEN_MODEL} from './hashState'
-import FolderOpenIcon from '@mui/icons-material/FolderOpen'
+import CreateNewFolderIcon from '@mui/icons-material/CreateNewFolderOutlined'
 
 
 /**
@@ -44,7 +44,7 @@ export default function OpenModelControl() {
   return (
     <ControlButtonWithHashState
       title='Open'
-      icon={<FolderOpenIcon className='icon-share'/>}
+      icon={<CreateNewFolderIcon className='icon-share'/>}
       isDialogDisplayed={isOpenModelVisible}
       setIsDialogDisplayed={setIsOpenModelVisible}
       hashPrefix={HASH_PREFIX_OPEN_MODEL}
