+++ src/Components/Open/OpenModelControl.test.jsx
@@ -14,10 +14,8 @@ describe('Open Model Dialog', () => {
     const {getByTestId, getByText} = render(<OpenModelControlFixture/>)
     const openControlButton = getByTestId('control-button-open')
     fireEvent.click(openControlButton)
-    const projectsTab = getByText('Projects')
-    fireEvent.click(projectsTab)
     const loginTextMatcher = (content, node) => {
-      const hasText = (_node) => _node.textContent.includes('Host your IFC models on GitHub and log in to BLDRS')
+      const hasText = (_node) => _node.textContent.includes('Please login to GitHub')
       const nodeHasText = hasText(node)
       const childrenDontHaveText = Array.from(node.children).every(
           (child) => !hasText(child),
@@ -32,11 +30,9 @@ describe('Open Model Dialog', () => {
 
   it('Renders file selector if the user is logged in', async () => {
     mockedUseAuth0.mockReturnValue(mockedUserLoggedIn)
-    const {getByTestId, getByText} = render(<OpenModelControlFixture/>)
+    const {getByTestId} = render(<OpenModelControlFixture/>)
     const openControlButton = getByTestId('control-button-open')
     fireEvent.click(openControlButton)
-    const projectsTab = getByText('Projects')
-    fireEvent.click(projectsTab)
     const File = getByTestId('openFile')
     const Repository = await getByTestId('openRepository')
     expect(File).toBeInTheDocument()
