+++ src/Components/Tabs.jsx
@@ -10,16 +10,16 @@ import {assertDefined} from '../utils/assert'
  * @property {boolean} [isScrollable] Enable scrolling for many (> 5) tabs
  * @return {React.Component}
  */
-export default function Tabs({tabLabels, currentTab, actionCb, isScrollable = false}) {
+export default function Tabs({tabLabels, actionCb, isScrollable = false}) {
   assertDefined(tabLabels, actionCb)
-  const [value, setValue] = useState(currentTab)
+  const [value, setValue] = useState(0)
   const handleChange = (event, newValue) => {
     setValue(newValue)
     actionCb(newValue)
   }
   return (
     <MuiTabs value={value} onChange={handleChange} centered variant={isScrollable ? 'scrollable' : 'fullWidth'}>
-      {tabLabels.map((tab) => <Tab key={tab} label={tab} data-testid={tab}/>)}
+      {tabLabels.map((tab) => <Tab key={tab} label={tab}/>)}
     </MuiTabs>
   )
 }
