+++ src/store/OpenSlice.js
@@ -9,7 +9,5 @@ export default function createOpenSlice(set, get) {
   return {
     isOpenEnabled: true,
     setIsOpenEnabled: (is) => set(() => ({isOpenEnabled: is})),
-    currentTab: 0,
-    setCurrentTab: (currentTab) => set(() => ({currentTab: currentTab})),
   }
 }
